First draft of API Needs
Went through GitHub API for health indicators we have discussed thus
far. These commit is the first pass at data we would like to aggregate
and perform pattern detection on. I view this as an initial “survey”
set of data that we would sample 100 projects, for example, and
compare, contrast & cluster.